Split unit tests from `filter.rs` into `filter_tests.rs`
